creating a cart app
