+++ requirements.txt
@@ -1,6 +1,6 @@
 django==4.2.11
 djangorestframework==3.15.1
-psycopg2==2.8.6
+psycopg2==2.9.9
 drf-yasg
 djangorestframework-simplejwt
 coverage
