+++ shop/cart/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from cart.models import Cart
+
+admin.site.register(Cart)
