+++ shop/shop/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Category, Product, Customer, Specification, Cart
+from .models import Category, Product, Customer, Specification
 
 
 @admin.register(Category)
@@ -14,4 +14,3 @@ class ProductAdmin(admin.ModelAdmin):
 
 admin.site.register(Customer)
 admin.site.register(Specification)
-admin.site.register(Cart)
