+++ shop/shop/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import Category, Specification, Product, CartProduct, Cart
+from shop.models import Category, Specification, Product
 
 
 class CategorySerializer(serializers.ModelSerializer):
@@ -22,21 +22,3 @@ class ProductSerializer(serializers.ModelSerializer):
     class Meta:
         model = Product
         fields = ('id', 'name', 'slug', 'price', 'discounted_price', 'remaining_product', 'specifications', 'category')
-
-
-class CartProductSerializer(serializers.ModelSerializer):
-    product = serializers.StringRelatedField()
-    user = serializers.StringRelatedField()
-
-    class Meta:
-        model = CartProduct
-        fields = ('qty', 'user', 'product', )
-
-
-class CartSerializer(serializers.ModelSerializer):
-    owner = serializers.StringRelatedField()
-    products = CartProductSerializer(many=True)
-
-    class Meta:
-        model = Cart
-        fields = '__all__'
